// components/OrderDetailDialog.tsx

import { Button } from "@/components/ui/button";
import { Dialog, DialogContent, DialogHeader, DialogTitle } from "@/components/ui/dialog";
import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from "@/components/ui/table";
import { MY_ORDERS_QUERY_RESULT } from "@/sanity.types";
import { urlFor } from "@/sanity/lib/image";
import Link from "next/link";
import Image from "next/image";

import React from "react";
import PriceFormatter from "@/components/PriceFormatter";

interface OrderDetailsDialogProps {
    order: MY_ORDERS_QUERY_RESULT[number] | null;
    isOpen: boolean;
    onClose: () => void;
}

const OrderDetailDialog: React.FC<OrderDetailsDialogProps> = ({
    order,
    isOpen,
    onClose,
}) => {
    if (!order) return null;
    return (
        <Dialog open={isOpen} onOpenChange={onClose}>
            <DialogContent className="!max-w-4xl w-full max-h-[90vh] overflow-y-auto">
                <DialogHeader>
                    <DialogTitle>
                        Chi tiết đặt hàng - {order?.orderNumber}
                    </DialogTitle>
                </DialogHeader>
                <div className="mt-4">
                    <p>
                        <strong>Khách hàng:</strong> {order?.customerName}
                    </p>
                    <p>
                        <strong>Email:</strong> {order?.email}
                    </p>
                    <p>
                        <strong>Ngày đặt:</strong>{" "}
                        {new Date(order._createdAt).toLocaleDateString("vi-VN")}
                    </p>
                    <p>
                        <strong>Trạng thái:</strong>{" "}
                        <span className="capitalize text-green-600 font-medium">
                            {order?.status}
                        </span>
                    </p>
                    <p>
                        <strong>Số hóa đơn:</strong> {order?.invoice?.number}
                    </p>
                    {order?.invoice && (
                        <Button className="bg-transparent border text-darkColor/80 mt-2 hover:text-darkColor hover:border-darkColor hover:bg-darkColor/10 hoverEffect">
                            {order?.invoice?.hosted_invoice_url && (
                                <Link href={order?.invoice?.hosted_invoice_url} target="_blank">
                                    In hoá đơn
                                </Link>
                            )}
                        </Button>
                    )}
                </div>
                <Table>
                    <TableHeader>
                        <TableRow>
                            <TableHead>Product Name</TableHead>
                            <TableHead>Quantity</TableHead>
                            <TableHead>Price</TableHead>
                        </TableRow>
                    </TableHeader>
                    <TableBody>
                        {order.products?.map((product, index) => (
                            <TableRow key={index}>
                                <TableCell className="flex items-center gap-2 ">
                                    {product?.product?.images && (
                                        <Image
                                            src={urlFor(product?.product?.images[0]).url()}
                                            alt="productImage"
                                            width={50}
                                            height={50}
                                            className="border rounded-sm"
                                        />
                                    )}
                                    {product?.product && product?.product?.name}
                                </TableCell>
                                <TableCell className="">{product?.quantity}</TableCell>
                                <TableCell>
                                    <PriceFormatter
                                        amount={product?.product?.price}
                                        className="text-black font-medium"
                                    />
                                </TableCell>
                            </TableRow>
                        ))}
                    </TableBody>
                </Table>
                <div className="mt-4 text-right flex items-center justify-end">
                    <div className="w-44 flex flex-col gap-1">
                        {order?.amountDiscount !== 0 && (
                            <div className="w-full flex items-center justify-between">
                                <strong>Discount:</strong>
                                <PriceFormatter
                                    amount={order?.amountDiscount}
                                    className="text-black font-bold"
                                />
                            </div>
                        )}

                        {order?.amountDiscount !== 0 && (
                            <div className="w-full flex items-center justify-between">
                                <strong>Tạm tính:</strong>
                                <PriceFormatter
                                    amount={(order?.totalPrice as number) + (order?.amountDiscount as number)}
                                    className="text-black font-bold"
                                />
                            </div>
                        )}
                        <div className="w-full flex items-center justify-between">
                            <strong>Tổng tiền:</strong>
                            <PriceFormatter
                                amount={order?.totalPrice}
                                className="text-black font-bold"
                            />
                        </div>
                    </div>
                </div>
            </DialogContent>
        </Dialog>
    )
};

export default OrderDetailDialog;
